TryLoadFromFilePaths - Improve exception handling (Split Unit Test)

+++ tests/NLog.UnitTests/LogFactoryTests.cs
@@ -95,29 +95,75 @@ namespace NLog.UnitTests
         }
 
         [Fact]
-        public void InaccessibleNLogConfigFileDoesNotThrow()
+        public void Configuration_InaccessibleNLog_doesNotThrowException()
         {
-            var tempDirectory = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString("N"));
-            var configFile = Path.Combine(tempDirectory, "NLog.config");
+            string tempDirectory = null;
 
             try
             {
-                Directory.CreateDirectory(tempDirectory);
-                File.WriteAllText(configFile, "<nlog />");
-                LogFactory logFactory = new LogFactory();
-                logFactory.SetCandidateConfigFilePaths(new[] { configFile });
-                using (var fileLock = new FileStream(configFile, FileMode.Open, FileAccess.ReadWrite, FileShare.None))
+                // Arrange
+                var logFactory = CreateEmptyNLogFile(out tempDirectory, out var configFile);
+                using (OpenStream(configFile))
                 {
+                    // Act
                     var loggingConfig = logFactory.Configuration;
+
+                    // Assert
                     Assert.Null(loggingConfig);
-                    Assert.ThrowsAny<Exception>(() => logFactory.LoadConfiguration(configFile));
                 }
+
+                // Assert
+                Assert.NotNull(logFactory.Configuration);
             }
             finally
             {
-                if (Directory.Exists(tempDirectory))
+                if (tempDirectory != null && Directory.Exists(tempDirectory))
                     Directory.Delete(tempDirectory, true);
             }
+        }
+
+        [Fact]
+        public void LoadConfiguration_InaccessibleNLog_throwException()
+        {
+            string tempDirectory = null;
+
+            try
+            {
+                // Arrange
+                var logFactory = CreateEmptyNLogFile(out tempDirectory, out var configFile);
+                using (OpenStream(configFile))
+                {
+                    // Act
+                    var ex = Record.Exception(() => logFactory.LoadConfiguration(configFile));
+
+                    // Assert
+                    Assert.IsType<IOException>(ex);
+                }
+
+                // Assert
+                Assert.NotNull(logFactory.LoadConfiguration(configFile).Configuration);
+            }
+            finally
+            {
+                if (tempDirectory != null && Directory.Exists(tempDirectory))
+                    Directory.Delete(tempDirectory, true);
+            }
+        }
+
+        private static FileStream OpenStream(string configFile)
+        {
+            return new FileStream(configFile, FileMode.Open, FileAccess.ReadWrite, FileShare.None);
+        }
+
+        private static LogFactory CreateEmptyNLogFile(out string tempDirectory, out string filePath)
+        {
+            tempDirectory = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString("N"));
+            filePath = Path.Combine(tempDirectory, "NLog.config");
+            Directory.CreateDirectory(tempDirectory);
+            File.WriteAllText(filePath, "<nlog />");
+            LogFactory logFactory = new LogFactory();
+            logFactory.SetCandidateConfigFilePaths(new[] { filePath });
+            return logFactory;
         }
 
         [Fact]
